Move current token to token::legacy
